further clean up in QueryPatternUtils
